CON-739 Replace message box popups with jQuery UI Dialog CON-640 Update of third party libraries

+++ contenido/includes/include.grouprights_menu.php
@@ -6,18 +6,14 @@
  * Description:
  * CONTENIDO Groups Menu
  *
- * Requirements:
- * @con_php_req 5.0
- *
- *
- * @package    CONTENIDO Backend Includes
- * @version    1.1.2
- * @author     Timo A. Hummel
- * @copyright  four for business AG <www.4fb.de>
- * @license    http://www.contenido.org/license/LIZENZ.txt
- * @link       http://www.4fb.de
- * @link       http://www.contenido.org
- * @since      file available since CONTENIDO release <= 4.6
+ * @package CONTENIDO Backend Includes
+ * @version 1.1.2
+ * @author Timo A. Hummel
+ * @copyright four for business AG <www.4fb.de>
+ * @license http://www.contenido.org/license/LIZENZ.txt
+ * @link http://www.4fb.de
+ * @link http://www.contenido.org
+ * @since file available since CONTENIDO release <= 4.6
  */
 
 if (!defined('CON_FRAMEWORK')) {
@@ -29,42 +25,33 @@ $tpl->set('s', 'SID', $sess->id);
 
 if (($action == "group_delete") && ($perm->have_perm_area_action($area, $action))) {
 
-   $sql = "DELETE FROM "
-             .$cfg["tab"]["groups"].
-          " WHERE
-             group_id = \"" .cSecurity::escapeDB($groupid, $db)."\"";
-   $db->query($sql);
+    $sql = "DELETE FROM " . $cfg["tab"]["groups"] . " WHERE
+             group_id = \"" . cSecurity::escapeDB($groupid, $db) . "\"";
+    $db->query($sql);
 
-   $sql = "DELETE FROM ".
-               $cfg["tab"]["groupmembers"].
-               " WHERE group_id = \"" .cSecurity::escapeDB($groupid, $db)."\"";
+    $sql = "DELETE FROM " . $cfg["tab"]["groupmembers"] . " WHERE group_id = \"" . cSecurity::escapeDB($groupid, $db) . "\"";
 
-   $db->query($sql);
+    $db->query($sql);
 
+    $sql = "DELETE FROM " . $cfg["tab"]["rights"] . " WHERE user_id = \"" . cSecurity::escapeDB($groupid, $db) . "\"";
 
-   $sql = "DELETE FROM ".
-               $cfg["tab"]["rights"].
-               " WHERE user_id = \"" .cSecurity::escapeDB($groupid, $db)."\"";
-
-   $db->query($sql);
+    $db->query($sql);
 }
 
 $sql = "SELECT
             groupname, group_id, perms
         FROM
-            ".$cfg["tab"]["groups"]."
+            " . $cfg["tab"]["groups"] . "
         ORDER BY
             groupname ASC";
 
-
-if ($restriction == 1)
-{
-     $sql = "SELECT
+if ($restriction == 1) {
+    $sql = "SELECT
             A.groupname AS groupname, A.group_id as group_id, A.perms as perms
         FROM
-            ".$cfg["tab"]["groups"]." AS A,
-            ".$cfg["tab"]["rights"]." AS B,
-            ".$cfg["tab"]["actions"]." AS C
+            " . $cfg["tab"]["groups"] . " AS A,
+            " . $cfg["tab"]["rights"] . " AS B,
+            " . $cfg["tab"]["actions"] . " AS C
         WHERE
             C.name = 'front_allow' AND
             B.user_id = A.group_id AND
@@ -76,14 +63,13 @@ if ($restriction == 1)
             groupname ASC";
 }
 
-if ($restriction == 3)
-{
-     $sql = "SELECT
+if ($restriction == 3) {
+    $sql = "SELECT
             A.groupname AS groupname, A.group_id as group_id, A.perms as perms
         FROM
-            ".$cfg["tab"]["groups"]." AS A,
-            ".$cfg["tab"]["rights"]." AS B,
-            ".$cfg["tab"]["actions"]." AS C
+            " . $cfg["tab"]["groups"] . " AS A,
+            " . $cfg["tab"]["rights"] . " AS B,
+            " . $cfg["tab"]["actions"] . " AS C
         WHERE
             C.name NOT LIKE 'front_allow' AND
             B.user_id = A.group_id AND
@@ -103,41 +89,33 @@ $thisperm = explode(',', $auth->auth['perm']);
 
 $accessibleClients = $classclient->getAccessibleClients();
 
-while ($db->next_record())
-{
+while ($db->next_record()) {
 
     $groupperm = explode(',', $db->f('perms'));
 
     $allow = false;
 
     // Sysadmin check
-    if (in_array("sysadmin", $thisperm))
-    {
+    if (in_array("sysadmin", $thisperm)) {
         $allow = true;
     }
 
     // Admin check
-    foreach ($accessibleClients as $key => $value)
-    {
-        if (in_array("client[".$key."]", $groupperm))
-        {
+    foreach ($accessibleClients as $key => $value) {
+        if (in_array("client[" . $key . "]", $groupperm)) {
             $allow = true;
         }
     }
 
     // Group check
-    foreach ($groupperm as $localperm)
-    {
+    foreach ($groupperm as $localperm) {
 
-        if (in_array($localperm, $thisperm))
-        {
+        if (in_array($localperm, $thisperm)) {
             $allow = true;
         }
-
     }
 
-    if ($allow == true)
-    {
+    if ($allow == true) {
 
         $groupid = $db->f("group_id");
         $groupname = $db->f("groupname");
@@ -146,24 +124,16 @@ while ($db->next_record())
 
         $tmp_mstr = '<a href="javascript:conMultiLink(\'%s\', \'%s\', \'%s\', \'%s\')">%s</a>';
         $area = "groups";
-        $mstr = sprintf($tmp_mstr, 'right_top',
-                                       $sess->url("main.php?area=$area&frame=3&groupid=$groupid"),
-                                       'right_bottom',
-                                       $sess->url("main.php?area=groups_overview&frame=4&groupid=$groupid"),
-                                       $groupname);
+        $mstr = sprintf($tmp_mstr, 'right_top', $sess->url("main.php?area=$area&frame=3&groupid=$groupid"), 'right_bottom', $sess->url("main.php?area=groups_overview&frame=4&groupid=$groupid"), $groupname);
 
-        $mstr2 = sprintf($tmp_mstr, 'right_top',
-                                       $sess->url("main.php?area=$area&frame=3&groupid=$groupid"),
-                                       'right_bottom',
-                                       $sess->url("main.php?area=groups_overview&frame=4&groupid=$groupid"),
-                                       '<img src="images/spacer.gif" border="0" width="15">');
+        $mstr2 = sprintf($tmp_mstr, 'right_top', $sess->url("main.php?area=$area&frame=3&groupid=$groupid"), 'right_bottom', $sess->url("main.php?area=groups_overview&frame=4&groupid=$groupid"), '<img src="images/spacer.gif" border="0" width="15">');
 
-        if ($perm->have_perm_area_action('groups',"groups_delete") ) {
-                $message = sprintf(i18n("Do you really want to delete the group %s?"),htmlspecialchars($groupname));
-                $deletebutton = "<a onClick=\"event.cancelBubble=true;check=confirm('".$message."'); if (check==true) { location.href='".$sess->url("main.php?area=groups&action=group_delete&frame=$frame&groupid=$groupid&del=")."#deletethis'};\" href=\"#\"><img src=\"".$cfg['path']['images']."delete.gif\" border=\"0\" width=\"13\" height=\"13\" alt=\"".i18n("Delete group")."\" title=\"".i18n("Delete group")."\"></a>";
-            } else {
-                $deletebutton = "";
-            }
+        if ($perm->have_perm_area_action('groups', "groups_delete")) {
+            $message = sprintf(i18n("Do you really want to delete the group %s?"), htmlspecialchars($groupname));
+            $deletebutton = "<a onClick=\"event.cancelBubble=true;check=confirm('" . $message . "'); if (check==true) { location.href='" . $sess->url("main.php?area=groups&action=group_delete&frame=$frame&groupid=$groupid&del=") . "#deletethis'};\" href=\"#\"><img src=\"" . $cfg['path']['images'] . "delete.gif\" border=\"0\" width=\"13\" height=\"13\" alt=\"" . i18n("Delete group") . "\" title=\"" . i18n("Delete group") . "\"></a>";
+        } else {
+            $deletebutton = "";
+        }
 
         if ($_GET['groupid'] == $groupid) {
             $tpl->set('d', 'ID_MARKED', 'marked');
@@ -174,12 +144,11 @@ while ($db->next_record())
         $tpl->set('d', 'TEXT', $mstr);
         $tpl->set('d', 'ICON', $mstr2);
 
-        if ($perm->have_perm_area_action('groups',"groups_delete") ) {
-         $delTitle = i18n("Delete group");
-        $delDescr = sprintf(i18n("Do you really want to delete the following group:<br><br>%s<br>"),htmlspecialchars($groupname));
+        if ($perm->have_perm_area_action('groups', "groups_delete")) {
+            $delTitle = i18n("Delete group");
+            $delDescr = sprintf(i18n("Do you really want to delete the following group:<br><br>%s<br>"), htmlspecialchars($groupname));
 
-
-        $tpl->set('d', 'DELETE', '<a title="'.$delTitle.'" href="javascript://" onclick="box.confirm(\''.$delTitle.'\', \''.$delDescr.'\', \'deleteGroup(\\\''.$groupid.'\\\')\')"><img src="'.$cfg['path']['images'].'delete.gif" border="0" title="'.$delTitle.'" alt="'.$delTitle.'"></a>');
+            $tpl->set('d', 'DELETE', '<a title="' . $delTitle . '" href="javascript:void(0)" onclick="showConfirmation(&quot;' . $delDescr . '&quot;, function() { deleteGroup(' . $groupid . '); });return false;"><img src="' . $cfg['path']['images'] . 'delete.gif" border="0" title="' . $delTitle . '" alt="' . $delTitle . '"></a>');
         } else {
             $tpl->set('d', 'DELETE', '&nbsp;');
         }
@@ -188,5 +157,5 @@ while ($db->next_record())
     }
 }
 
-# Generate template
+// Generate template
 $tpl->generate($cfg['path']['templates'] . $cfg['templates']['grouprights_menu']);
